Move localscanner to certrefresh
